+++ README.md
@@ -1,3 +1,7 @@
 
+Python astronomy data & results e-something
+===========================================
+
+
 functions and helpers for displaying radio astronomy datasets in an ipython notebook
